improved the Progress and add the SoundBtn
