+++ src/ui-components/Progress/Progress.module.less
@@ -1,0 +1,85 @@
+.container{
+    width: 50%;
+    text-align: center;
+    position: relative;
+    // background-color: rgb(217, 236, 236);
+}
+
+
+.progressContainer{
+    margin: 0px auto;
+    width: 80%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 30px;
+    position: relative;
+    z-index: 3;
+}
+
+
+.progressContainer::before{
+    content: '';
+    width: 100%;
+    height: 4px;
+    background-color: rgb(194, 191, 191);
+    position: absolute;
+    top: 50%;
+    left: 0px;
+    transform: translateY(-50%);
+    z-index: -1;
+    /* transition: all 0.3s ease; */
+}
+
+.progressBar{
+    width: 33px;
+    height: 4px;
+    background-color: rgb(54, 88, 199);
+    position: absolute;
+    top: 50%;
+    left: 0px;
+    transform: translateY(-50%);
+    z-index: 0;
+    transition: all 0.3s ease;
+}
+
+.circle{
+    width: 33px;
+    line-height: 30px;
+    border-radius: 50%;
+    border: 3px solid rgb(199, 199, 199);
+    background-color: white;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: #999;
+    z-index: 2;
+    transition: all 0.4s ease-in;
+}
+.circle.active{
+    border-color: rgb(54, 88, 199);
+}
+
+.btn{
+    width: 80px;
+    line-height: 30px;
+    background-color: rgb(12, 114, 248);
+    border: none;
+    border-radius: 10px;
+    cursor: pointer;
+    transition: all 0.3s;
+    color: white;
+    margin-left: 5px;
+    cursor: pointer;
+
+}
+.btn:disabled{
+    background-color: rgb(196, 189, 189);
+    cursor: default;
+}
+.btn:disabled:active{
+    transform: scale(1);
+}
+.btn:active{
+    transform: scale(0.90);
+}
