+++ src/ui-components/SoundBoard/SoundBtn.module.less
@@ -1,0 +1,29 @@
+.container{
+    width: 100%;
+    padding: 20px;
+    height: 100%;
+    background-color: rgb(200, 151, 240);
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    #buttons{
+        width: 50%;
+        line-height: 50px;
+        display: flex;
+        justify-content: space-around;
+    }
+    .btn{
+        background-color: rgb(119, 68, 160);
+        color: rgb(216, 205, 205);
+        width: 15%;
+        line-height: 40px;
+        border: none;
+        cursor: pointer;
+        font-size: 18px;
+        border-radius: 4px;
+        overflow: hidden;
+    }
+    .btn:active{
+        opacity: 0.8;
+    }
+}
