+++ src/ui-components/index.ts
@@ -1,6 +1,10 @@
 import Button from "./Button/Button";
 import SwithCard from "./SwithCards/SwithCard";
+import Progress from "./Progress/Progress";
+import SoundBtn from "./SoundBoard/SoundBtn";
 export {
     Button,
-    SwithCard
+    SwithCard,
+    Progress,
+    SoundBtn
 }
